Вынес свойства для полей ввода в отдельный класс
